LPD-79027 Extract
